# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para la gestión de una peluquería",
        "title": "Peluquería API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/cliente": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve una lista de todos los clientes registrados.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener todos los clientes",
                "responses": {
                    "200": {
                        "description": "Lista de clientes encontrados",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dtos.GetClientDto"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo cliente en el sistema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear cliente",
                "parameters": [
                    {
                        "description": "Datos del cliente",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.ClientDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente creado exitosamente",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Datos inválidos. Ejemplo: El cliente ya existe",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cliente/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve los datos de un cliente específico.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener cliente por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dtos.GetClientDto"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza los datos de un cliente específico.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar cliente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados del cliente",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.ClientDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente actualizado exitosamente",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Datos o ID inválidos",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Elimina un cliente específico.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar cliente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado exitosamente",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.ClientAppointmentDto": {
            "type": "object",
            "properties": {
                "appointment_date": {
                    "type": "string",
                    "example": "30/09/2002 16:30"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string",
                    "example": "finalizado"
                }
            }
        },
        "dtos.ClientDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@gmail.com"
                },
                "last_name": {
                    "type": "string",
                    "example": "Garcia Mendez"
                },
                "name": {
                    "type": "string",
                    "example": "Valentino"
                },
                "phone": {
                    "type": "string",
                    "example": "343534345"
                }
            }
        },
...
        "dtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Detalle del error",
                    "type": "string",
                    "example": "Descripción del error"
                },
                "status": {
                    "description": "Siempre \"error\"",
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "dtos.GetClientDto": {
            "type": "object",
            "properties": {
                "appointments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.ClientAppointmentDto"
                    }
                },
                "email": {
                    "type": "string",
                    "example": "example@gmail.com"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string",
                    "example": "Garcia Mendez"
                },
                "name": {
                    "type": "string",
                    "example": "Valentino"
                },
                "phone": {
                    "type": "string",
                    "example": "343534345"
                }
            }
        },
...
        "dtos.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Datos adicionales"
                },
                "message": {
                    "description": "Mensaje descriptivo",
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "status": {
                    "description": "Ejemplo de estado para éxito o error",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    }
}